third_party/WebKit/LayoutTests/bluetooth/getCharacteristics/device-disconnects-during.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return setBluetoothFakeAdapter('DisconnectingHeartRateAdapter')
    .then(() => requestDeviceWithKeyDown({
      filters: [{services: ['heart_rate']}],
      optionalServices: [request_disconnection_service_uuid]
    }))
    .then(device => device.gatt.connect())
    .then(gattServer => {
      let heart_rate_service;
      return gattServer.getPrimaryService('heart_rate')
        .then(hrs => heart_rate_service = hrs)
        .then(() => gattServer.getPrimaryService(
          request_disconnection_service_uuid))
        .then(service => service.getCharacteristic(
          request_disconnection_characteristic_uuid))
        .then(requestDisconnection => {
          requestDisconnection.writeValue(new Uint8Array([0]));
          return assert_promise_rejects_with_message(
            heart_rate_service.getCharacteristics(),
            new DOMException(
              'GATT Server disconnected while retrieving characteristics.',
              'NetworkError'));
        });
    });
}, 'Device disconnects during getCharacteristics. Reject with NetworkError.');
</script>
